app/requirements.txt
Django==4.1.2
pytz==2022.4
tzdata==2022.4
django-admin-interface==0.22.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.4
django-ninja==0.22.2


